Feature/mem buffer

+++ include_private/jsdrv_prv/js220_i128.h
@@ -47,8 +47,8 @@ js220_i128 js220_i128_udiv(js220_i128 dividend, uint64_t divisor, uint64_t * rem
 js220_i128 js220_i128_lshift(js220_i128 x, int32_t shift);
 js220_i128 js220_i128_rshift(js220_i128 x, int32_t shift);
 double js220_i128_to_f64(js220_i128 x, uint32_t q);
-double js220_i128_compute_std(int64_t x1, js220_i128 x2, uint32_t n, uint32_t q);
-js220_i128 js220_i128_compute_integral(js220_i128 x, uint32_t n);
+double js220_i128_compute_std(int64_t x1, js220_i128 x2, uint64_t n, uint32_t q);
+js220_i128 js220_i128_compute_integral(js220_i128 x, uint64_t n);
 
 JSDRV_CPP_GUARD_END
